+++ iclub-app/src/main/java/za/co/iclub/pss/trans/IclubAccountTypeTrans.java
@@ -8,7 +8,7 @@ public class IclubAccountTypeTrans {
 	
 	public static IclubAccountTypeBean fromWStoUI(IclubAccountTypeModel model) {
 		IclubAccountTypeBean bean = new IclubAccountTypeBean();
-		bean.setAtId(model.getAtId().longValue());
+		bean.setAtId(model.getAtId());
 		bean.setAtLongDesc(model.getAtLongDesc());
 		bean.setAtShortDesc(model.getAtShortDesc());
 		bean.setAtStatus(model.getAtStatus());
@@ -17,7 +17,7 @@ public class IclubAccountTypeTrans {
 	
 	public static IclubAccountTypeModel fromUItoWS(IclubAccountTypeBean bean) {
 		IclubAccountTypeModel model = new IclubAccountTypeModel();
-		model.setAtId(bean.getAtId().longValue());
+		model.setAtId(bean.getAtId());
 		model.setAtLongDesc(bean.getAtLongDesc());
 		model.setAtShortDesc(bean.getAtShortDesc());
 		model.setAtStatus(bean.getAtStatus());
@@ -26,7 +26,7 @@ public class IclubAccountTypeTrans {
 	
 	public static IclubAccountTypeModel fromORMtoWS(IclubAccountType bean) {
 		IclubAccountTypeModel model = new IclubAccountTypeModel();
-		model.setAtId(bean.getAtId().longValue());
+		model.setAtId(bean.getAtId());
 		model.setAtLongDesc(bean.getAtLongDesc());
 		model.setAtShortDesc(bean.getAtShortDesc());
 		model.setAtStatus(bean.getAtStatus());
